fix bug

+++ main.js
@@ -24,16 +24,26 @@ app.get('/', function(req, res) {
   res.sendfile('index.html');
 });
 
-
+    if (fs.existsSync('public/app/info/m2ain.js')) {
+    // Do something
+        console.log(1)
+}else{
+    console.log(0)
+}
 app.post('/newProject',function(req,res){
     var text = req.body.text;
-    var mainJs = 'function initApp' + text + '() {\n var app = document.getElementById("application-window");\n app.src = "app/' + text + '/index.html";}\n\n createApp({\n    nameApp: "' + text + '",\n    func: "initApp' + text + '();"\n });'
-    fs.mkdir('public/app/' + text);
-    fs.mkdir('public/app/' + text + '/css');
-    fs.mkdir('public/app/' + text + '/js');
-    fs.writeFile('public/app/' + text + '/index.html', '<p>Hello, world !</p>', function(e){});
-    fs.writeFile('public/app/' + text + '/main.js', mainJs, function(e){});
-    res.end("OK");
+    
+    if(fs.existsSync('public/app/' + text + '.js')) {
+        console.log('Приложение не может быть создано !'.red)
+    } else {
+        var mainJs = 'function initApp' + text + '() {\n var app = document.getElementById("application-window");\n app.src = "app/' + text + '/index.html";}\n\n createApp({\n    nameApp: "' + text + '",\n    func: "initApp' + text + '();"\n });'
+        fs.mkdir('public/app/' + text);
+        fs.mkdir('public/app/' + text + '/css');
+        fs.mkdir('public/app/' + text + '/js');
+        fs.writeFile('public/app/' + text + '/index.html', '<p>Hello, world !</p>', function(e){});
+        fs.writeFile('public/app/' + text + '.js', mainJs, function(e){});
+        res.end("OK");
+    }
 });
 
 app.listen(8081);
